requirements.txt
redis
kafka-python
six
clickhouse-connect
